+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - `cli/aux/toggle`: Fix memory leaks
 - `cli/add`: Fix overlapping memory regions
 - `cli`: Fix alias check for `cd` commands
+- `cli/pgn`: Separate hook and pgn module
 
 ### Deleted